requirements.txt
aniso8601==1.2.0
Babel==2.3.4
blinker==1.4
click==6.7
decorator==4.0.10
Flask==0.12
Flask-Admin==1.4.2
Flask-Babel==0.11.1
Flask-BabelEx==0.9.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.5
Flask-WTF==0.14
itsdangerous==0.24
Jinja2==2.9.3
MarkupSafe==0.23
olefile==0.44
Pillow==4.0.0
pkg-resources==0.0.0
psycopg2==2.6.2
pyasn1==0.1.9
python-dateutil==2.6.0
python3-ldap==0.9.8.4
pytz==2016.10
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.4
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.15
WTForms==2.1
